lovejoy/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
wtforms==3.1.2
email_validator==2.2.0
passlib[argon2]==1.7.4
pyotp==2.9.0
itsdangerous==2.2.0
bleach==6.1.0
Pillow==10.4.0
qrcode==7.4.2
python-dotenv==1.0.1

flask==3.0.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
wtforms==3.1.2
email_validator==2.2.0
argon2-cffi==23.1.0
pyotp==2.9.0
itsdangerous==2.2.0
bleach==6.1.0
Pillow==10.4.0
qrcode==7.4.2
python-dotenv==1.0.1